+++ CHANGES.rst
@@ -2,8 +2,8 @@ Changelog
 =========
 
 
-1.8.3 (unreleased)
-------------------
+2.0 (unreleased)
+----------------
 
 - Add Plone 6 compatibility, drop Plone 4 & Plone 5 compatibility
   [laulaz]